Avoid implicit conversion

+++ src/software/firmware/includes/pressure_utl.h
@@ -15,7 +15,7 @@
  * @param sensorValue Value read from the analog input connected to the sensor
  * @return the pressure in mmH2O
  */
-int16_t convertSensor2Pressure(uint16_t sensorValue);
+int16_t convertSensor2Pressure(uint32_t sensorValue);
 
 /**
  * @brief Reset the value of void filteredRawPressure to 0.
